Fix an LLVM IR issue with calling a yielding function via a trait.
Prior to this commit, the internal code that decides how to generate
LLVM IR for a program was broken by not recognizing that the function
could yield unless it had a yield in the body - a trait function
need not have a body, and thus could trigger wrong LLVM IR generation.

After this fix, a `:yields` declaration is enough to be noted that
the function could yield, even if it has no body. This fixes the issue.

This commit has no added regression test, but the next commit will add code
that exercises this path, so I'm not worried about it right now.