Fix an issue with yielding call LLVM IR generation.
Prior to this commit, there was a compiler crash when
trying to generate an abstract yielding call that had
no reachable concrete implementation.

This was happening in cases where we have possibly-reachable
`TraceData.Observer` use sites in a program where there is
no actual reachable concrete type that is a `TraceData.Observer`.

This commit also adds some additional troubleshooting information
to be printed when an unexpected exception is raised during
the generation of a function declaration or implementation.