Fix erroneous "double-dot" output in `FloatingPoint.Format`.
Prior to this commit, a number like `0.5`
would be formatted with two dots, as: `0..5`

This commit fixes that issue by adding an additonal condition
for printing the 2nd dot.

This commit also removes a no-longer-used dependency of
`FloatingPoint.Format` types on `Integer.Format.Decimal`.
Apparently they referenced that type earlier in their
implementation, but they no longer do, so I've removed the reference.