Properly check for manifests in parent directory
Sorry, I forgot to mention this change in our previous discussion.  The current behavior does not check the parent directory for manifest files.
For example, if the program is run from directory "/foo/bar/", `try_dirname` starts by being set to "/foo/bar/", and gets set to "/foo/" in line 259.  Then, it is checked for starting with "/foo/bar/" in line 260.  Since this is false, the loop only executes once.
With this change, `try_dirname` starts by being set to "/foo/bar/", gets set to "/foo/", and gets checked for starting with "/foo/".  The loop continues, `try_dirname` gets set to "/" and gets checked for starting with "/foo/".  Then, the loop ends.