DEPRECATED: `_until` methods using boolean-driven loop interruption.
Using a boolean yield block result to determine whether an "each"-like
loop should continue is an old pattern I added to Savi early on.

Now this pattern is superceded by the fact that the `break` macro now
works with yield blocks just as well as it works with inline loop macros.

So the old pattern of defining a separate `_until` method is now
kind of an anti-pattern, and shouldn't be used anymore, going forward.

This commit marks those standard-library methods as being `DEPRECATED`.
They'll be removed in the future.