Fix a memory safety bug with error values.
Prior to this commit, the new error values feature had some
edge cases which had tests (in the language semantics specs),
but the failure of those tests was hidden by the fact that
`savi run` was silently swallowing the SEGV signal and
not setting a non-zero exit code on the parent process.

Once that `savi run` issue was fixed, the `try_spec` tests
had to be commented out because some of them were causing
segmentation faily (SEGV signal).

This commit restores the commented-out tests and fixes the
underlying code generation issues that were causing the SEGV signal.