Remove half-baked lambda concept from the language.
This lambda construct has never been thought through fully,
has never been used in real Savi code, and is a serious thorn
in the side of the compiler architecture (just as the corresponding
lambda mechanism is in the Pony compiler).

This was just something I was experimenting with during the
very early days of compiler development. Not anything finished.

We will add some form of lambda mechanism in the future, but it
will be different enough from this mechanism that there isn't
any point in keeping it around cluttering the compiler.