Fix networking setup in language runtime.
In order for networking stuff to work as desired, we need to
pass "true" for `init_network` in the `pony_language_features_init_t`
passed to the `pony_start` call that starts the runtime.

Without that, things like the `TCP` library won't work properly.
(for example, on UNIX systems, a `SIGPIPE` may interrupt the
program when writing to a TCP socket whose other ends is known
by the kernel to have already closed).